feat(translation): add language flags as svg
